requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.1.0
cachetools==6.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
curl_cffi==0.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
mexc-api==0.0.2
multidict==6.1.0
openpyxl==3.1.5
propcache==0.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymexc==1.2.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
tradingview-ta==3.3.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0
wsaccel==0.6.7
yarl==1.18.3
